# Progress on Ass2
Finished the implementation of the advanced feature of auto query expansion,
the program should run if my logic is correct, however usually there needs debugging.

Waiting for:
Kanru to finish the modified version of invoke arguments reading.

To Kanru:
please implement your code in this version of code base that I have done, and then give it a test and we can fix bugs together.
time is running out so hopefully we can get it done quickly.